updating the parser information
