+++ opengrok-tools/src/main/python/opengrok_tools/java.py
@@ -25,7 +25,8 @@
 import argparse
 import sys
 
-from .utils.java import Java, get_javaparser
+from .utils.java import Java
+from opengrok_tools.utils.parsers import get_javaparser
 from .utils.log import get_console_logger, get_class_basename, print_exc_exit
