+++ opengrok-tools/src/main/python/opengrok_tools/mirror.py
@@ -36,18 +36,19 @@ import re
 import sys
 import tempfile
 from logging.handlers import RotatingFileHandler
+
 from filelock import Timeout, FileLock
 
+from opengrok_tools.utils.parsers import get_baseparser
+from .scm.repository import RepositoryException
 from .utils.hook import run_hook
 from .utils.log import get_console_logger, get_class_basename, \
-    add_log_level_argument, print_exc_exit
+    print_exc_exit
 from .utils.opengrok import get_repos, get_config_value, get_repo_type
 from .utils.readconfig import read_config
 from .utils.repofactory import get_repository
-from .utils.utils import is_exe, check_create_dir, get_int, diff_list,\
+from .utils.utils import is_exe, check_create_dir, get_int, diff_list, \
     is_web_uri
-from .scm.repository import RepositoryException
-
 
 major_version = sys.version_info[0]
 if major_version < 3:
@@ -118,10 +119,10 @@ def get_repos_for_project(logger, project, ignored_repos, **kwargs):
 def main():
     ret = 0
 
-    parser = argparse.ArgumentParser(description='project mirroring')
+    parser = argparse.ArgumentParser(description='project mirroring',
+                                     parents=[get_baseparser()])
 
     parser.add_argument('project')
-    add_log_level_argument(parser)
     parser.add_argument('-c', '--config',
                         help='config file in JSON/YAML format')
     parser.add_argument('-U', '--uri', default='http://localhost:8080/source',
