+++ opengrok-tools/src/main/python/opengrok_tools/utils/java.py
@@ -25,9 +25,7 @@
 import platform
 from .command import Command
 from .utils import is_exe
-from .log import add_log_level_argument
 import os
-import argparse
 
 
 class Java(Command):
@@ -108,24 +106,3 @@ class Java(Command):
                     java = os.path.join(java_home, 'bin', 'java')
 
         return java
-
-
-def get_javaparser():
-    parser = argparse.ArgumentParser(add_help=False)
-    add_log_level_argument(parser)
-    parser.add_argument('-j', '--java',
-                        help='path to java binary')
-    parser.add_argument('-J', '--java_opts',
-                        help='java options', action='append')
-    parser.add_argument('-e', '--environment', action='append',
-                        help='Environment variables in the form of name=value')
-
-    group = parser.add_mutually_exclusive_group(required=True)
-    group.add_argument('-a', '--jar',
-                       help='Path to jar archive to run')
-    group.add_argument('-c', '--classpath',
-                       help='Class path')
-
-    parser.add_argument('options', nargs='+', help='options')
-
-    return parser
